requirements.txt
azure-core
azure-search-documents
tqdm